dom/network/tests/test_networkstats_enabled_perm.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test to ensure NetworkStats is not accessible when it is disabled</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">
<script type="application/javascript">

"use strict";

SimpleTest.waitForExplicitFinish();

// Test to ensure NetworkStats is not accessible when it is disabled.
SpecialPowers.pushPrefEnv({'set': [["dom.mozNetworkStats.enabled", true]]},
                          function() {
  SpecialPowers.pushPermissions([{ 'type': 'networkstats-manage', 'allow': 1, 'context': window.document }],
                                function() {

    ok(SpecialPowers.hasPermission("networkstats-manage", document),
       "Has permission 'networkstats-manage'.");

    ok(SpecialPowers.getBoolPref("dom.mozNetworkStats.enabled"),
       "Preference 'dom.mozNetworkStats.enabled' is true.");

    ok('mozNetworkStats' in navigator, "navigator.mozNetworkStats should exist");

    ok(navigator.mozNetworkStats instanceof MozNetworkStatsManager,
       "navigator.mozNetworkStats should be a MozNetworkStatsManager object");

    SimpleTest.finish();
  });
});

</script>
</pre>
</body>
</html>
